altera a nomenclatura dos modulos para melhor descricao
